refactor(app): extract useModelPicker from session-route
Model picker modal state (open/compact/query/options), the new-providers
toast event + pending-localStorage triggers, the lazy option loader with
'Recently added' flagging, and the org-restriction filter move verbatim to
domains/session/modals/use-model-picker.ts. The hook self-acquires the
desktop-restriction context and exposes raw setters so call sites stay
one-line renames. The render watchdog moved below the hook to keep its
modelPickerOpen attribution field.

session-route.tsx: 2,934 -> 2,815 lines (3,328 at phase start).
settings-route still carries its sibling copy of the option loader; adopting
this hook there is queued as its own commit.

Tested: pnpm typecheck clean; bun test 58 pass; pnpm build clean;
pnpm test:e2e 23/23 assertions ok.
